Refactor copy-pasted section into its own component
