+++ src/module/vue/components/character-sheet-tabs/sf-progresses.vue
@@ -1,180 +1,7 @@
 <template>
-  <div class="flexcol">
-    <div class="flexcol ironsworn__drop__target" data-drop-type="progress">
-      <transition-group name="slide" tag="div" class="nogrow">
-        <div
-          class="flexrow nogrow"
-          v-for="(item, i) in activeItems"
-          :key="item._id"
-        >
-          <order-buttons
-            v-if="editMode"
-            :i="i"
-            :length="activeItems.length"
-            @sortUp="sortUp"
-            @sortDown="sortDown"
-          />
-          <progress-box
-            :item="item"
-            :actor="actor"
-            :showStar="true"
-            @completed="progressCompleted"
-          />
-        </div>
-      </transition-group>
-      <progress-controls foeCompendium="starforgedencounters" />
-    </div>
-
-    <div class="item-row nogrow progress-completed" style="margin-top: 1rem">
-      <!-- TODO: refactor this as a component in PR for collapsible/progressive disclosure element -->
-      <h3>
-        <btn-faicon
-          :disabled="completedItems.length === 0"
-          class="text collapse-control"
-          :class="completedClass"
-          :icon="completedCaret"
-          @click="data.expandCompleted = !data.expandCompleted"
-          >{{ $t('IRONSWORN.Completed') }}</btn-faicon
-        >
-      </h3>
-      <transition
-        name="slide"
-        tag="div"
-        class="nogrow completed"
-        style="margin: 0; padding: 0"
-      >
-        <div v-if="data.expandCompleted">
-          <transition-group name="slide" tag="div" class="nogrow">
-            <div
-              class="flexrow"
-              v-for="(item, i) in completedItems"
-              :key="item._id"
-            >
-              <order-buttons
-                v-if="editMode"
-                :i="i"
-                :length="completedItems.length"
-                @sortUp="completedSortUp"
-                @sortDown="completedSortDown"
-              />
-              <progress-box :item="item" :actor="actor" :showStar="true" />
-            </div>
-          </transition-group>
-        </div>
-      </transition>
-    </div>
-  </div>
+  <active-completed-progresses />
 </template>
 
-<style lang="less">
-.progress-completed {
-  .collapse-control {
-    text-transform: uppercase;
-    height: inherit;
-    width: inherit;
-    flex-grow: 1;
-  }
-  h3 {
-    display: flex;
-  }
-  .highlighted {
-    background-color: lightyellow;
-  }
-}
-</style>
-
-<style lang="less" scoped>
-h3 {
-  margin: 5px 0;
-  transition: background-color 0.2s ease;
-  i {
-    width: 15px;
-    text-align: center;
-  }
-
-  &.highlighted {
-    background-color: lightyellow;
-  }
-}
-.slide-enter-active,
-.slide-leave-active {
-  max-height: 106px;
-  &.completed {
-    max-height: 400px;
-  }
-}
-</style>
-
 <script setup lang="ts">
-import { computed, inject, reactive, Ref } from 'vue'
-import { $ActorKey } from '../../provisions'
-import OrderButtons from '../order-buttons.vue'
-import ProgressBox from '../progress/progress-box.vue'
-import ProgressControls from '../progress-controls.vue'
-import BtnFaicon from '../buttons/btn-faicon.vue'
-
-const data = reactive({
-  expandCompleted: false,
-  highlightCompleted: false,
-})
-
-const actor = inject('actor') as Ref
-const $actor = inject($ActorKey)
-
-const progressItems = computed(() => {
-  return actor.value.items
-    .filter((x) => x.type === 'progress')
-    .filter((x) => x.data.subtype !== 'bond')
-    .sort((a, b) => (a.sort || 0) - (b.sort || 0))
-})
-const activeItems = computed(() => {
-  return progressItems.value.filter((x) => !x.data.completed)
-})
-const completedItems = computed(() => {
-  return progressItems.value.filter((x) => x.data.completed)
-})
-const editMode = computed(() => {
-  return actor.value.flags['foundry-ironsworn']?.['edit-mode']
-})
-const completedCaret = computed(() => {
-  return data.expandCompleted ? 'caret-down' : 'caret-right'
-})
-const completedClass = computed(() => {
-  return data.highlightCompleted ? 'highlighted' : undefined
-})
-
-let highlightCompletedTimer: NodeJS.Timer | undefined
-function progressCompleted() {
-  data.highlightCompleted = true
-  clearTimeout(highlightCompletedTimer)
-  highlightCompletedTimer = setTimeout(() => {
-    data.highlightCompleted = false
-  }, 2000)
-}
-
-async function applySort(oldI, newI, sortBefore, filterFn) {
-  const foundryItems = $actor?.items
-    .filter((x) => x.type === 'progress')
-    .filter((x) => x.data.data.subtype !== 'bond')
-    .filter(filterFn)
-    .sort((a, b) => (a.data.sort || 0) - (b.data.sort || 0))
-  const updates = SortingHelpers.performIntegerSort(foundryItems[oldI], {
-    target: (foundryItems ?? [])[newI],
-    siblings: foundryItems,
-    sortBefore,
-  })
-  await Promise.all(updates.map(({ target, update }) => target.update(update)))
-}
-function sortUp(i) {
-  applySort(i, i - 1, true, (x) => !x.data.data.completed)
-}
-function sortDown(i) {
-  applySort(i, i + 1, false, (x) => !x.data.data.completed)
-}
-function completedSortUp(i) {
-  applySort(i, i - 1, true, (x) => x.data.data.completed)
-}
-function completedSortDown(i) {
-  applySort(i, i + 1, false, (x) => x.data.data.completed)
-}
+import ActiveCompletedProgresses from '../active-completed-progresses.vue'
 </script>
